+++ src/main/java/kitchenpos/order/domain/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
 import static kitchenpos.exception.ExceptionType.INVALID_PRICE_EXCEPTION;
