+++ src/main/java/kitchenpos/order/ui/request/OrderRequest.java
@@ -1,9 +1,9 @@
-package kitchenpos.ui.dto.request;
+package kitchenpos.order.ui.request;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderRequest {
     private Long orderTableId;
