+++ src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.exception.CustomIllegalArgumentException;
-import kitchenpos.dao.JpaOrderRepository;
+import kitchenpos.order.domain.JpaOrderRepository;
 import kitchenpos.table.domain.JpaOrderTableRepository;
 import kitchenpos.table.domain.JpaTableGroupRepository;
 import kitchenpos.domain.Order;
