+++ src/test/java/kitchenpos/application/HappyOrderServiceTest.java
@@ -8,8 +8,8 @@ import kitchenpos.domain.Order;
 import kitchenpos.ui.dto.request.MenuGroupRequest;
 import kitchenpos.menu.ui.request.MenuProductRequest;
 import kitchenpos.menu.ui.request.MenuRequest;
-import kitchenpos.ui.dto.request.OrderLineItemRequest;
-import kitchenpos.ui.dto.request.OrderRequest;
+import kitchenpos.order.ui.request.OrderLineItemRequest;
+import kitchenpos.order.ui.request.OrderRequest;
 import kitchenpos.table.ui.request.OrderTableRequest;
 import kitchenpos.product.ui.request.ProductRequest;
 import org.assertj.core.api.Assertions;
