+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 import kitchenpos.dao.JpaMenuGroupRepository;
 import kitchenpos.menu.domain.JpaMenuRepository;
-import kitchenpos.dao.JpaOrderRepository;
+import kitchenpos.order.domain.JpaOrderRepository;
 import kitchenpos.table.domain.JpaOrderTableRepository;
 import kitchenpos.product.domain.JpaProductRepository;
 import kitchenpos.table.domain.JpaTableGroupRepository;
@@ -23,8 +23,8 @@ import kitchenpos.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.product.domain.Product;
 import kitchenpos.table.domain.TableGroup;
-import kitchenpos.ui.dto.request.OrderLineItemRequest;
-import kitchenpos.ui.dto.request.OrderRequest;
+import kitchenpos.order.ui.request.OrderLineItemRequest;
+import kitchenpos.order.ui.request.OrderRequest;
 import kitchenpos.table.ui.request.OrderTableIdRequest;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.SpringBootTest;
